redhat: Improve python check
RH-Author: Miroslav Rezanina <mrezanin@redhat.com>
Message-id: <1538054579-30273-1-git-send-email-mrezanin@redhat.com>
Patchwork-id: 82304
O-Subject: [RHEL8/rhel qemu-kvm PATCH] redhat: Improve python check
Bugzilla:
RH-Acked-by: Wei Huang <wei@redhat.com>
RH-Acked-by: Danilo de Paula <ddepaula@redhat.com>
RH-Acked-by: Wainer dos Santos Moschetta <wainersm@redhat.com>

From: Miroslav Rezanina <mrezanin@redhat.com>

BZ: n/a
Upstream: n/a
Brew: n/a
Branch: rhel8/master-2.12.0

We introduce check for python interpreter to properly run brew builds
on RHEL 8 systems. However, this check is not working on Fedora so
rh-brew don't work there.

Changing interpreter check to work properly on any host.

+++ redhat/scripts/process-patches.sh
@@ -29,11 +29,19 @@ SRPMDIR=rpmbuild/SRPM
 BUILDDIR=${SOURCES}/${BUILDDIR_NAME}
 SPEC=rpmbuild/SPECS/${SPECNAME}
 
-if test "$(rpm --eval '%{?rhel}')" = 7; then
-  LOCAL_PYTHON=/usr/bin/python
-else
-  LOCAL_PYTHON=$(rpm --eval '%{__python3}')
+LOCAL_PYTHON=$(
+  if python --version > /dev/null; then
+    echo python
+  elif python3 --version > /dev/null; then
+    echo python3
+  fi 2> /dev/null
+)
+
+if [ -z "$LOCAL_PYTHON" ]; then
+  echo "No python interpreter found"
+  exit 1
 fi
+echo "Using $LOCAL_PYTHON"
 
 # Pre-cleaning
 rm -rf .tmp asection psection patchlist
